app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.121.3",
  "uvicorn>=0.38.0",
  "pydantic>=2.12.4",
  "boto3>=1.41.1",
  "duckdb",
  "pytz>=2025.2",
  "aiosqlite>=0.21.0",
  "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["package*"]